Move HookService to Hooks\ namespace
