+++ Classes/Interfaces/Hook.php
@@ -3,7 +3,7 @@
 namespace OliverKlee\Seminars\Interfaces;
 
 /**
- * Extend this interface to use Your interface with the HookService
+ * Extend this interface to use Your interface with the Hooks\HookService
  *
  * @author Michael Kramer <m.kramer@mxp.de>
  */
